Add methods to get and create Minecraft dev pack paths
Added `GetPreviewDevelopmentBehaviourPackPath` and
`GetPreviewDevelopmentResourcePackPath` methods to the `Addon`
class in `Addon.cpp` and `Addon.hpp`. These methods return the
paths to the preview development behavior and resource packs
using the `LOCALAPPDATA` environment variable. If the directories
do not exist, the methods create them. Paths are stored in static
`std::string` variables for efficiency.